Initial kin-sim, with cubic spline interp
This interp isn't very stable as can be seen by spline-stable, when the
times get really long relative to the velocities on each side. Basically,
the system is too "heavy", so the variables overshoot by a lot. This is
just inherent to the solution, and higher-order polynomials (at the min
distance solution) surprisingly aren't better.

Instead, we want to move to a solution with predictable overshoot.