New kinematics (Jacobian + inertia calculation) + ESC control attempt
Tried to control the Hobbywing Quicrun ESC as a speed-torque controller,
but it wouldn't hold position properly, instead trying to advance when it
stalled. This led to the migration to a dumbe 3-phase driver.

Also new kinematics have landed: Jacobian (more specifically inverse Jacobian
plus a bunch of routines from CLAPACK) and the fwd kinematics for inertia
calculations, also originally for tuning the control algorithm for J0
(may become obsolete if the control is sufficiently robust on the independent
motor driver)