Fix kinematics and producer-consumer buffering
Inverse kinematics had a couple of calculation bugs that are detailed in
`ctrl/ctrl_test.c`.

Producer-consumer had lots of problems with choreographing buffer
freshness, consumer empty and advancing the starting point for the
interpolation. All of the flags were restructured to avoid the
multitude of races that were causing sudden jumps in the trajectory.